Phase 4: split settings route into per-section components
